Change add, update and delete methods due the tables and logis of working the web app

+++ data_manager/sqlite_data_manager.py
@@ -4,8 +4,8 @@ from data_manager.data_models import db, Movie, User, UserMovie
 class SQLiteDataManager(DataManagerInterface):
     def __init__(self, app):
         self.app = app
-        self.db = db # sqlalchemy objecr from data_models
-        db.init_app(app) # inizialisation of the db_init
+        self.db = db # sqlalchemy object from data_models
+        db.init_app(app) # inizialisation of the db in the app
         with app.app_context():
             self.db.create_all() # create all the tables
     
@@ -24,17 +24,28 @@ class SQLiteDataManager(DataManagerInterface):
         self.db.session.add(user_data)
         self.db.commit()
     
-    def add_movie(self, movie):
+    def add_movie(self, movie, user_id, watchlist_status, user_rating):
         movie_data = Movie(**movie)
         self.db.session.add(movie_data)
         self.db.commit()
-    
-    def update_movie(self, movie):
-        updated_movie = Movie(**movie)
-        self.db.session.add(updated_movie)
+        movie_id = movie_data.id
+        user_movie_data = UserMovie(user_id=user_id, movie_id=movie_id, watchlist_status=watchlist_status, user_rating=user_rating)
+        self.db.session.add(user_movie_data)
         self.db.commit()
     
-    def delete_movie(self, movie_id):
-        movie_to_delete = self.db.session.query(Movie).filter(Movie.id == movie_id).first()
-        self.db.session.delete(movie_to_delete)
-        self.db.commit()
+    def update_movie(self, user_id, movie_id, rating, status):
+        user_movie = self.db.session.query(UserMovie).filter(UserMovie.movie_id == movie_id, UserMovie.user_id == user_id).first
+        if user_movie:
+            user_movie.watchlist_status = status
+            user_movie.user_rating = rating
+            db.session.commit()
+            return True  # Indicate success
+        return False  # Indicate failure (movie not found)
+    
+    def delete_movie(self, user_id, movie_id):
+        movie_to_delete = self.db.session.query(UserMovie).filter(UserMovie.movie_id == movie_id, UserMovie.user_id == user_id).first
+        if movie_to_delete:
+            self.db.session.delete(movie_to_delete)
+            self.db.commit()
+            return True
+        return False
